# puller: optimize spanFrontier