<!--
Thank you for contributing to TiDB-CDC! Please read MD's [CONTRIBUTING](https://github.com/pingcap/tidb-cdc/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Fixes #461.

The main issue is the bad performance of `spanFrontier.Forward` when the input `Span` overlaps many tracked span. Here are the benchmark results on `master`.

```
BenchmarkSpanFrontier/5k-32               633535              1839 ns/op              16 B/op          1 allocs/op
BenchmarkSpanFrontier/10k-32              600888              1976 ns/op              16 B/op          1 allocs/op
BenchmarkSpanFrontier/50k-32              428908              2408 ns/op              16 B/op          1 allocs/op
BenchmarkSpanFrontier/100k-32             446808              2691 ns/op              16 B/op          1 allocs/op
BenchmarkSpanFrontierOverlap/5k_5-32               40912             28493 ns/op            2919 B/op         33 allocs/op
BenchmarkSpanFrontierOverlap/5k_10-32              22344             53121 ns/op            5779 B/op         55 allocs/op
BenchmarkSpanFrontierOverlap/5k_100-32              2394            456998 ns/op           52856 B/op        421 allocs/op
BenchmarkSpanFrontierOverlap/5k_500-32               514           2201803 ns/op          243704 B/op       2026 allocs/op
BenchmarkSpanFrontierOverlap/10k_5-32              40291             30066 ns/op            2919 B/op         33 allocs/op
BenchmarkSpanFrontierOverlap/10k_10-32             20901             56005 ns/op            5781 B/op         55 allocs/op
BenchmarkSpanFrontierOverlap/10k_100-32             2547            495105 ns/op           52856 B/op        422 allocs/op
BenchmarkSpanFrontierOverlap/10k_500-32              501           2319103 ns/op          243704 B/op       2025 allocs/op
BenchmarkSpanFrontierOverlap/50k_5-32              34308             33785 ns/op            2920 B/op         33 allocs/op
BenchmarkSpanFrontierOverlap/50k_10-32             19089             64235 ns/op            5784 B/op         55 allocs/op
BenchmarkSpanFrontierOverlap/50k_100-32             1938            577198 ns/op           52856 B/op        421 allocs/op
BenchmarkSpanFrontierOverlap/50k_500-32              429           2677039 ns/op          243704 B/op       2025 allocs/op
BenchmarkSpanFrontierOverlap/100k_5-32             32064             35909 ns/op            2920 B/op         33 allocs/op
BenchmarkSpanFrontierOverlap/100k_10-32            17402             67018 ns/op            5784 B/op         55 allocs/op
BenchmarkSpanFrontierOverlap/100k_100-32            1758            624967 ns/op           52856 B/op        421 allocs/op
BenchmarkSpanFrontierOverlap/100k_500-32             412           2880261 ns/op          243704 B/op       2025 allocs/op
```

### What is changed and how it works?

#### Use skip list to replace the interval tree
The main problem here is the cost of the interval tree. The performance is even worse due to we will always delete overlapped spans and insert them back to index.

https://github.com/pingcap/ticdc/blob/0193ea137c9b2fd361feedd5b0440e36edc82e24/cdc/puller/span_frontier.go#L223-L239

Let's rethink what `Forward` really do:
1. Query all overlapped spans;
1. For each overlapped spans
	* If the span is fully covered by the input range, we will try to increase it's `ts`;
	* Otherwise, we will create two spans, one of them will inherit the old `ts` and another will use the new `ts`. I'll call this operation as `split` below.

It should be noted that in the second step, we will never create `Span` to inherit the `ts` of the input range.

https://github.com/pingcap/ticdc/blob/0193ea137c9b2fd361feedd5b0440e36edc82e24/cdc/puller/span_frontier.go#L204-L213

For example, we have `1, 3`, `5, 7` and `9, 12` in the index, and the input is `2, 11`.
we will do *split* operation on `1, 3` and `9, 12` and never create `3, 5` and `7, 9`. As you can see, we will never create two overlapped spans in the index.

So the actual steps of `Froward` are:
1. Find the first overlapped span, and check do we need to *split* a left sub-span;
2. Iterate to the next span, and update it's ts if needed;
3. If there is no next overlapped span, check do we need to *split* a right sub-span.

A skip list can do much better than the interval tree, it is easy to implement, more efficient to do iteration. So I use a skip list to index the `Start` of each `Span`. Here is the benchmark results after this optimization:

```
BenchmarkSpanFrontier/5k-32              1648172               698 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/10k-32             1605222               773 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/50k-32             1297162               892 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/100k-32            1176512              1000 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_5-32             1000000              1060 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_10-32             764371              1404 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_100-32            132060              9124 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_500-32             27001             45481 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_5-32             987498              1094 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_10-32            821604              1473 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_100-32           129451              9406 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_500-32            26852             44615 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_5-32             990517              1248 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_10-32            740077              1655 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_100-32           128402              9462 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_500-32            27310             45109 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_5-32            865586              1369 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_10-32           656137              1755 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_100-32          117265             10120 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_500-32           24775             46282 ns/op               0 B/op          0 allocs/op
```

But it seems not good enough in some extreme scenarios.

#### Use Fibonacci heap to maintain the minimum `ts`
Now the bottleneck is we need to increase the `ts` of fully covered spans. It is not trivial to do such an increase key operation in a minimum binary heap, so we just remove and insert it again, 2 * x * O(log N), `x` is covered spans and `N` is the total spans in the index. 

But do we really need to maintain the states of the heap when we increase someone's `ts`? I think it is not necessary to do this, we actually just need to fix the heap before the next `Frontier` call, or even further, we just need to do this when the minimum `ts` is increased.

The Fibonacci heap's deferred fix is very suitable for this scenario, we just need to cut some nodes to root list, and defer the consolidate of the root list when actually need.

With this optimization, we can see that the performance has been greatly improved in these scenarios

```
BenchmarkSpanFrontier/5k-32              1823144               650 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/10k-32             1562533               754 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/50k-32             1436476               838 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontier/100k-32            1340686               893 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_5-32             1651561               727 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_10-32            1412175               897 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_100-32            487171              2634 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/5k_500-32            119612             10282 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_5-32            1499804               822 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_10-32           1308764               889 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_100-32           560109              2734 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/10k_500-32           130642             10574 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_5-32            1239646               972 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_10-32           1258894               988 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_100-32           488380              2842 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/50k_500-32           128559             10859 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_5-32           1232728               991 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_10-32          1231414               996 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_100-32          546308              2848 ns/op               0 B/op          0 allocs/op
BenchmarkSpanFrontierOverlap/100k_500-32          126040              9883 ns/op               0 B/op          0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test