@amyangfei 2020-04-23
## prepare environment

$ sysbench --config-file=config oltp_insert --create_secondary=off --tables=16 --table-size=100000 prepare

split region
```
for i in $(seq 1 16); do
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (0) and (11000000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (11000000) and (12100000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (12100000) and (13200000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (13200000) and (14300000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (14300000) and (15400000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (15400000) and (16500000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (16500000) and (17600000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (17600000) and (18700000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (18700000) and (19800000) regions 1000"
        mysql -h 172.16.5.113 -u root -P 4000 -e "split table cdc_bench.sbtest$i between (19800000) and (20900000) regions 1000"
done
```

region count: 159877

pd-ctl scheduler add -u http://172.16.5.113:2379 random-merge-scheduler

merge-schedule-limit = 64


|                                                              | CPU        | Forward percent | Region count |
| ------------------------------------------------------------ | ---------- | --------------- | ------------ |
| master                                                       | 960%       | 10.69%          | 159877       |
| [pr#479](https://github.com/pingcap/ticdc/pull/479)          | 920%       | 6.38%           | 159877       |
| random merge, master                                         | 940%-1000% | 20.39%          | 159877 -> 37 |
| random merge, [pr#479](https://github.com/pingcap/ticdc/pull/479) | 960%-1100% | 0.22%           | 159877 -> 37 |

<img width="677" alt="Screen Shot 2020-04-23 at 12 01 13" src="https://user-images.githubusercontent.com/1527315/80082901-8f54cd00-8587-11ea-9716-0321e4de1f6a.png">


merge-schedule-limit = 128


|                                                              | CPU  | Forward percent | Region count | pprof file                                                |
| ------------------------------------------------------------ | ---- | --------------- | ------------ | --------------------------------------------------------- |
| [pr#480](https://github.com/pingcap/ticdc/pull/480)          | 155% | 81.33%          | 159877 -> 37 | [cpu profile](https://transfer.sh/4bE1H/pprof.cdc.samples.cpu.038.pb.gz) |
| [pr#479](https://github.com/pingcap/ticdc/pull/479) with [pr#480](https://github.com/pingcap/ticdc/pull/480) |      |                 | 159877->37   |                                                           |